+++ packages/react-pose/src/posed.tsx
@@ -6,6 +6,7 @@ import {
   PoseContextProps
 } from './components/PoseElement.types';
 import { DomPopmotionConfig } from 'popmotion-pose/lib/types';
+import { warning } from 'hey-listen';
 
 // TODO: Something is wrong with the TypeScript React bindings here,
 // I've had to typecast a component as a Component a couple times now.
@@ -16,13 +17,13 @@ export type ComponentFactory = (
 ) => (props: PoseElementProps) => ReactElement<any>;
 
 export type Posed = {
-  (): ComponentFactory;
+  (component: React.Component): ComponentFactory;
   [key: string]: ComponentFactory;
 };
 
-const componentCache = new Map<string, ComponentFactory>();
+const componentCache = new Map<string | React.Component, ComponentFactory>();
 
-const createComponentFactory = (key: string) => {
+const createComponentFactory = (key: string | React.Component) => {
   const componentFactory: ComponentFactory = (poseConfig = {}) => ({
     withParent = true,
     ...props
@@ -51,15 +52,32 @@ const createComponentFactory = (key: string) => {
   return componentFactory;
 };
 
-const getComponentFactory = (key: string) =>
+const getComponentFactory = (key: string | React.Component) =>
   componentCache.has(key)
     ? componentCache.get(key)
     : createComponentFactory(key);
 
-const posed: Posed = new Proxy(function() {} as Posed, {
-  get: (target, key: string) => getComponentFactory(key),
-  apply: (target, thisArg, argumentsList) =>
-    getComponentFactory(argumentsList[0])
-});
+let posed = ((component: React.Component) =>
+  getComponentFactory(component)) as Posed;
+
+if (typeof Proxy !== 'undefined') {
+  posed = new Proxy(posed, {
+    get: (target, key: string) => getComponentFactory(key),
+    apply: (target, thisArg, argumentsList) =>
+      getComponentFactory(argumentsList[0])
+  });
+
+  // If the browser doesn't support Proxy, offer a subselection of element types.
+  // This gets killed in 2019 yeah
+} else {
+  warning(
+    false,
+    `You're seeing this message because you're testing in a browser that doesn't support Proxy. If you need to support this browser, you can only use posed.div, posed.span, posed.a, posed.li and posed.button`
+  );
+  ['div', 'span', 'a', 'button', 'li'].reduce((acc, key) => {
+    acc[key] = getComponentFactory(key);
+    return acc;
+  }, posed);
+}
 
 export default posed;